+++ internal/type_wrapper.go
@@ -2,7 +2,6 @@ package internal
 
 // #include <memory.h>
 // #include <stdlib.h>
-// #include <stdbool.h>
 import "C"
 
 import (
@@ -16,21 +15,6 @@ import (
 // It just forces pointer/type reinterpretation with unsafe.Pointer.
 func ReinterpretCast[RET, SRC any](src SRC) RET {
 	return *(*RET)(unsafe.Pointer(&src))
-}
-
-// Number is a generic type for Go/C types that can be used as a number.
-// It could be anything that you can convert to that type (e.g. C.int is a Number,
-// because it can be directly converted to int)
-type Number interface {
-	~int | ~int8 | ~int16 | ~int32 | ~int64 |
-		~uint | ~uint8 | ~uint16 | ~uint32 | ~uint64 |
-		~float32 | ~float64 |
-		~bool // in C bool is technically a number
-}
-
-// WrapNumberPtr is a generic method to convert GOTYPE (int32/float32 e.t.c.) into CTYPE (c_int/c_float e.t.c.)
-func WrapNumberPtr[CTYPE Number, GOTYPE Number](goValue *GOTYPE) (wrapped *CTYPE, finisher func()) {
-	return ReinterpretCast[*CTYPE](goValue), func() {}
 }
 
 // WrapString converts Go string to C char*
@@ -77,61 +61,4 @@ func PtrToByteSlice(p unsafe.Pointer) []byte {
 
 func PtrToUint16Slice(p unsafe.Pointer) []uint16 {
 	return (*[unrealisticLargePointer / 2]uint16)(p)[:]
-}
-
-type StringBuffer struct {
-	ptr  unsafe.Pointer
-	size int
-}
-
-func NewStringBuffer(initialValue string) *StringBuffer {
-	rawText := []byte(initialValue)
-	bufSize := len(rawText) + 1
-	newPtr := C.malloc(C.size_t(bufSize))
-	zeroOffset := bufSize - 1
-	buf := PtrToByteSlice(newPtr)
-	copy(buf[:zeroOffset], rawText)
-	buf[zeroOffset] = 0
-
-	return &StringBuffer{ptr: newPtr, size: bufSize}
-}
-
-func (buf *StringBuffer) Free() {
-	C.free(buf.ptr)
-	buf.size = 0
-}
-
-func (buf *StringBuffer) ResizeTo(requestedSize int) {
-	bufSize := requestedSize
-	if bufSize < 1 {
-		bufSize = 1
-	}
-	newPtr := C.malloc(C.size_t(bufSize))
-	copySize := bufSize
-	if copySize > buf.size {
-		copySize = buf.size
-	}
-	if copySize > 0 {
-		C.memcpy(newPtr, buf.ptr, C.size_t(copySize))
-	}
-	PtrToByteSlice(newPtr)[bufSize-1] = 0
-	C.free(buf.ptr)
-	buf.ptr = newPtr
-	buf.size = bufSize
-}
-
-func (buf *StringBuffer) ToGo() string {
-	if (buf.ptr == nil) || (buf.size < 1) {
-		return ""
-	}
-	PtrToByteSlice(buf.ptr)[buf.size-1] = 0
-	return C.GoString((*C.char)(buf.ptr))
-}
-
-func (buf *StringBuffer) Ptr() unsafe.Pointer {
-	return buf.ptr
-}
-
-func (buf *StringBuffer) Size() int {
-	return buf.size
 }
